feat(utils): moving multisig utils into scripts
